internal: move to the top level so it can be used by cmd
Towads https://github.com/giantswarm/giantswarm/issues/4753.
