+++ pkg/compiler/compiler.go
@@ -11,7 +11,7 @@ import (
 	"github.com/giantswarm/micrologger"
 
 	"github.com/giantswarm/e2e-harness/pkg/harness"
-	"github.com/giantswarm/e2e-harness/pkg/internal/golang"
+	"github.com/giantswarm/e2e-harness/internal/golang"
 )
 
 type Config struct {
